+++ Demo/src/main/java/com/braintreepayments/demo/CardFragment.java
@@ -27,6 +27,7 @@ import com.braintreepayments.api.ThreeDSecureAdditionalInformation;
 import com.braintreepayments.api.ThreeDSecureClient;
 import com.braintreepayments.api.ThreeDSecureLauncher;
 import com.braintreepayments.api.ThreeDSecureNonce;
+import com.braintreepayments.api.ThreeDSecurePaymentAuthRequest;
 import com.braintreepayments.api.ThreeDSecurePostalAddress;
 import com.braintreepayments.api.ThreeDSecureRequest;
 import com.braintreepayments.api.ThreeDSecureResult;
@@ -35,6 +36,7 @@ import com.braintreepayments.api.ThreeDSecureV2LabelCustomization;
 import com.braintreepayments.api.ThreeDSecureV2TextBoxCustomization;
 import com.braintreepayments.api.ThreeDSecureV2ToolbarCustomization;
 import com.braintreepayments.api.ThreeDSecureV2UiCustomization;
+import com.braintreepayments.api.UserCanceledException;
 import com.braintreepayments.cardform.OnCardFormFieldFocusedListener;
 import com.braintreepayments.cardform.OnCardFormSubmitListener;
 import com.braintreepayments.cardform.utils.CardType;
@@ -89,8 +91,17 @@ public class CardFragment extends BaseFragment implements OnCardFormSubmitListen
         View view = inflater.inflate(R.layout.fragment_card, container, false);
 
         threeDSecureLauncher = new ThreeDSecureLauncher(this,
-                paymentAuthResult -> threeDSecureClient.tokenize(paymentAuthResult,
-                        this::handleThreeDSecureResult));
+                paymentAuthResult -> {
+                    threeDSecureClient.tokenize(paymentAuthResult, threeDSecureResult -> {
+                        if (threeDSecureResult instanceof ThreeDSecureResult.Success) {
+                            handlePaymentMethodNonceCreated(((ThreeDSecureResult.Success) threeDSecureResult).getNonce());
+                        } else if (threeDSecureResult instanceof ThreeDSecureResult.Failure) {
+                            handleError(((ThreeDSecureResult.Failure) threeDSecureResult).getError());
+                        } else if (threeDSecureResult instanceof ThreeDSecureResult.Cancel) {
+                            handleError(new UserCanceledException("User canceled 3DS."));
+                        }
+                    });
+                });
 
         cardForm = view.findViewById(R.id.card_form);
         cardForm.setOnFormFieldFocusedListener(this);
@@ -200,34 +211,26 @@ public class CardFragment extends BaseFragment implements OnCardFormSubmitListen
         autofillHelper.fillPostalCode("12345");
     }
 
-    private void handleThreeDSecureResult(ThreeDSecureResult threeDSecureResult, Exception error) {
-        safelyCloseLoadingView();
-        if (threeDSecureResult != null) {
-            ThreeDSecureNonce paymentMethodNonce = threeDSecureResult.getThreeDSecureNonce();
-            handlePaymentMethodNonceCreated(paymentMethodNonce);
-        } else {
-            handleError(error);
-        }
-    }
-
     private void handlePaymentMethodNonceCreated(PaymentMethodNonce paymentMethodNonce) {
         super.onPaymentMethodNonceCreated(paymentMethodNonce);
 
         final FragmentActivity activity = getActivity();
         if (!threeDSecureRequested && paymentMethodNonce instanceof CardNonce &&
-                Settings.isThreeDSecureEnabled(activity)) {
+                Settings.isThreeDSecureEnabled(activity) && !(paymentMethodNonce instanceof ThreeDSecureNonce)) {
             threeDSecureRequested = true;
             loading = ProgressDialog.show(activity, getString(R.string.loading),
                     getString(R.string.loading), true, false);
 
             ThreeDSecureRequest threeDSecureRequest = threeDSecureRequest(paymentMethodNonce);
             threeDSecureClient.createPaymentAuthRequest(requireContext(), threeDSecureRequest,
-                    (threeDSecureResult, error) -> {
-                        if (threeDSecureResult != null &&
-                                threeDSecureResult.getLookup().requiresUserAuthentication()) {
-                            threeDSecureLauncher.launch(threeDSecureResult);
-                        } else {
-                            handleThreeDSecureResult(threeDSecureResult, error);
+                    (paymentAuthRequest) -> {
+                        if (paymentAuthRequest instanceof ThreeDSecurePaymentAuthRequest.ReadyToLaunch) {
+                            threeDSecureLauncher.launch(
+                                    (ThreeDSecurePaymentAuthRequest.ReadyToLaunch) paymentAuthRequest);
+                        } else if (paymentAuthRequest instanceof ThreeDSecurePaymentAuthRequest.LaunchNotRequired) {
+                            handlePaymentMethodNonceCreated(((ThreeDSecurePaymentAuthRequest.LaunchNotRequired) paymentAuthRequest).getNonce());
+                        } else if (paymentAuthRequest instanceof ThreeDSecurePaymentAuthRequest.Failure) {
+                            handleError(((ThreeDSecurePaymentAuthRequest.Failure) paymentAuthRequest).getError());
                         }
                         safelyCloseLoadingView();
                     });